refactor: move accesibility managers in structure
