+++ lib/features/digital_guide/tabs/structure/business/ramps_accessibility_comments_manager.dart
@@ -1,8 +1,8 @@
 import "package:fast_immutable_collections/fast_immutable_collections.dart";
 
-import "../../../../../../l10n/app_localizations.dart";
-import "../../../../business/accessibility_comments_manager.dart";
-import "../models/ramp.dart";
+import "../../../../../l10n/app_localizations.dart";
+import "../../../business/accessibility_comments_manager.dart";
+import "../data/models/ramp.dart";
 
 class RampsAccessibilityCommentsManager extends AccessibilityCommentsManager {
   RampsAccessibilityCommentsManager({required this.ramps, required this.l10n});
