+++ lib/features/digital_guide/tabs/structure/business/stairway_accessibility_comments_manager.dart
@@ -1,8 +1,8 @@
 import "package:fast_immutable_collections/fast_immutable_collections.dart";
 
-import "../../../../../../l10n/app_localizations.dart";
-import "../../../../business/accessibility_comments_manager.dart";
-import "../models/stairway.dart";
+import "../../../../../l10n/app_localizations.dart";
+import "../../../business/accessibility_comments_manager.dart";
+import "../data/models/stairway.dart";
 
 class StairwayAccessibilityCommentsManager extends AccessibilityCommentsManager {
   StairwayAccessibilityCommentsManager({required this.stairway, required this.l10n});
