+++ lib/features/digital_guide/tabs/structure/presentation/views/ramps_view.dart
@@ -12,7 +12,7 @@ import "../../../../presentation/widgets/accessibility_button.dart";
 import "../../../../presentation/widgets/accessibility_profile_card.dart";
 import "../../../../presentation/widgets/bullet_list.dart";
 import "../../../../presentation/widgets/digital_guide_nav_link.dart";
-import "../../data/bussiness/ramps_accessibility_comments_manager.dart";
+import "../../business/ramps_accessibility_comments_manager.dart";
 import "../../data/models/ramp.dart";
 
 @RoutePage()
